move dialog to a separate class
