+++ skate-plugin/src/main/kotlin/com/slack/sgp/intellij/projectgen/ProjectGenMenuAction.kt
@@ -15,20 +15,15 @@
 */
 package com.slack.sgp.intellij.projectgen
 
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
-import com.intellij.openapi.diagnostic.logger
 import com.intellij.openapi.project.Project
-import com.slack.circuit.foundation.Circuit
-import com.slack.circuit.foundation.CircuitContent
-import com.slack.circuit.runtime.ui.ui
 import com.slack.sgp.intellij.tracing.SkateSpanBuilder
 import com.slack.sgp.intellij.tracing.SkateTraceReporter
-import java.io.File
-import java.nio.file.FileSystems
-import java.nio.file.Paths
+import com.slack.sgp.intellij.tracing.SkateTracingEvent
+import com.slack.sgp.intellij.tracing.SkateTracingEvent.EventType.PROJECT_GEN_OPENED
+import com.slack.sgp.intellij.util.isProjectGenMenuActionEnabled
+import com.slack.sgp.intellij.util.isTracingEnabled
 import java.time.Instant
 
 
@@ -42,46 +37,22 @@ constructor(
   private val startTimestamp = Instant.now()
 
   override fun actionPerformed(e: AnActionEvent) {
-    ProjectGenWindow(e.project).show()
+    val currentProject: Project = e.project ?: return
+    if (!currentProject.isProjectGenMenuActionEnabled()) return
+    ProjectGenWindow(currentProject).show()
+
+    if (currentProject.isTracingEnabled()) {
+      sendUsageTrace(currentProject)
+    }
   }
 
   fun sendUsageTrace(project: Project) {
+    skateSpanBuilder.addSpanTag("event", SkateTracingEvent(PROJECT_GEN_OPENED))
     SkateTraceReporter(project, offline)
       .createPluginUsageTraceAndSendTrace(
         "project_generator",
         startTimestamp,
         skateSpanBuilder.getKeyValueList()
       )
-  }
-
-  class ProjectGenWindow(private val currentProject: Project?) : ComposeDialog(currentProject) {
-    init {
-      title = "Project Generator"
-    }
-
-    @Composable
-    override fun dialogContent() {
-
-      logger<ProjectGenMenuAction>().info(currentProject?.basePath)
-      val rootDir = remember {
-        val path = currentProject?.basePath
-          ?: FileSystems.getDefault().getPath(".").toAbsolutePath().normalize().toFile().absolutePath
-        check(Paths.get(path).toFile().isDirectory) { "Must pass a valid directory" }
-        path
-      }
-      File(rootDir + "/.projectgenlock").createNewFile()
-
-      val circuit = remember {
-        Circuit.Builder()
-          .addPresenterFactory { _, _, _ -> ProjectGenPresenter(rootDir, ::doOKAction) }
-          .addUiFactory { _, _ ->
-            ui<ProjectGenScreen.State> { state, modifier -> ProjectGen(state, modifier) }
-          }
-          .build()
-      }
-      SlackDesktopTheme() {
-        CircuitContent(ProjectGenScreen, circuit = circuit)
-      }
-    }
   }
 }
